+++ src/M5000_emulator.h
@@ -1,1 +1,1 @@
-void M5000_emulator_init(uint8_t instance, int address);
+void M5000_emulator_init(uint8_t instance, uint8_t address);
