+++ src/harddisc_emulator.h
@@ -1,1 +1,1 @@
-void harddisc_emulator_init( uint8_t instance , int address);
+void harddisc_emulator_init( uint8_t instance , uint8_t address);
