+++ src/helpers.c
@@ -33,7 +33,7 @@ void helpers_bank_select(unsigned int gpio)
     }
 
 }
-void helpers_init( uint8_t instance , int address)
+void helpers_init( uint8_t instance , uint8_t address)
 {
    if (filesystemReadFile("6502code.bin",&helper_ram[0],sizeof(helper_ram)))
     {
