+++ src/helpers.h
@@ -1,1 +1,1 @@
-void helpers_init( uint8_t instance , int address);
+void helpers_init( uint8_t instance , uint8_t address);
